packages/opentelemetry-instrumentation-google-genai/pyproject.toml
[tool.coverage.run]
branch = true
source = ["opentelemetry/instrumentation/google_genai"]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.poetry]
name = "opentelemetry-instrumentation-google-genai"
version = "0.39.2"
description = "OpenTelemetry Google GenAI instrumentation"
authors = ["Gal Kleinman <gal@traceloop.com>", "Nir Gazit <nir@traceloop.com>"]
repository = "https://github.com/traceloop/openllmetry/tree/main/packages/opentelemetry-instrumentation-google-genai"
license = "Apache-2.0"
readme = "README.md"

[[tool.poetry.packages]]
include = "opentelemetry/instrumentation/google_genai"

[tool.poetry.dependencies]
python = ">=3.9,<4"
opentelemetry-api = "^1.28.0"
opentelemetry-instrumentation = ">=0.50b0"
opentelemetry-semantic-conventions = ">=0.50b0"
opentelemetry-semantic-conventions-ai = "0.4.3"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.2.0"
flake8 = "7.0.0"
pytest = "^8.2.2"
pytest-sugar = "1.0.0"
pytest-asyncio = "^0.26.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-sugar = "1.0.0"
vcrpy = "^6.0.1"
pytest-recording = "^0.13.1"
opentelemetry-sdk = "^1.27.0"
google-genai = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
instruments = ["google-genai"]

[tool.poetry.plugins."opentelemetry_instrumentor"]
google_genai = "opentelemetry.instrumentation.google_genai:GoogleGenAiSdkInstrumentor"
